게시글 작성 페이지 분리
