+++ src/pages/MainPage.vue
@@ -9,22 +9,8 @@
         </li>
       </ul>
       <hr />
-      <h2>게시글 작성</h2>
-      <form @submit.prevent="registerBoard">
-        <div>
-          <label for="title">제목:</label>
-          <input type="text" id="title" v-model="newBoard.title" required />
-        </div>
-        <div>
-          <label for="content">내용:</label>
-          <textarea id="content" v-model="newBoard.content" required></textarea>
-        </div>
-        <div>
-          <label for="writer">작성자:</label>
-          <input type="text" id="writer" v-model="newBoard.writer" required />
-        </div>
-        <button type="submit">작성</button>
-      </form>
+      <router-link to="/register">게시글 작성하기</router-link>
+      
     </div>
   </template>
   
@@ -59,18 +45,6 @@
           this.boards = response.data;
         } catch (error) {
           console.error('Error fetching boards:', error);
-        }
-      },
-      async registerBoard() {
-        try {
-          await api.post('/board/register', this.newBoard);
-          alert('게시글 작성 완료');
-          this.fetchBoards();
-          this.newBoard.title = '';
-          this.newBoard.content = '';
-          this.newBoard.writer = '';
-        } catch (error) {
-          console.error('작성 실패:', error);
         }
       }
     }
